+++ NEWS.md
@@ -168,7 +168,7 @@ matches for the query. Fixed. (#199 and #163)
 the token stream. Fixed.
 - An error when using `format()` on a `features_ngrams` object resulting in an
 error when using `knit_print()` on this object has been fixed (#200).
-
+- The `edit()` method can now be invoked on a `features` object (#165).
 
 # Documentation